moves segment_postprocess to modules
