+++ tests/modules/segment_postprocess/test_segment_postprocess.py
@@ -11,7 +11,7 @@ import sys
 sys.modules['suite2p'] = Mock()
 from ophys_etl.modules.suite2p_wrapper.schemas import \
         Suite2PWrapperSchema, Suite2PWrapperOutputSchema  # noqa: E402
-import ophys_etl.pipelines.segment_postprocess_pipeline as sbpipe  # noqa
+import ophys_etl.modules.segment_postprocess.__main__ as sbpipe  # noqa
 
 
 class MockSuite2PWrapper(argschema.ArgSchemaParser):
@@ -45,10 +45,10 @@ class MockPostProcess(argschema.ArgSchemaParser):
 
 @pytest.mark.suite2p_only
 @patch(
-        'ophys_etl.pipelines.segment_postprocess_pipeline.Suite2PWrapper',
+        'ophys_etl.modules.segment_postprocess.__main__.Suite2PWrapper',
         MockSuite2PWrapper)
 @patch(
-        'ophys_etl.pipelines.segment_postprocess_pipeline.PostProcessROIs',
+        'ophys_etl.modules.segment_postprocess.__main__.PostProcessROIs',
         MockPostProcess)
 def test_segment_postprocess_pipeline(tmp_path):
     """tests that satisfying the pipeline schema satisfies the
